+++ src/main/java/com/example/wbsp19devanshiganatraserverside/model/Module.java
@@ -4,23 +4,23 @@ import java.util.List;
 
 public class Module {
 
-    private Integer id;
+    private Long id;
     private String title;
     private List<Lesson> lessons;
 
     public Module () {}
 
-    public Module(Integer id, String title, List<Lesson> lessons) {
+    public Module(Long id, String title, List<Lesson> lessons) {
         this.id = id;
         this.title = title;
         this.lessons = lessons;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
